+++ CromIAM/src/test/scala/cromiam/webservice/EngineRouteSupportSpec.scala
@@ -1,7 +1,8 @@
 package cromiam.webservice
 
-import akka.http.scaladsl.model.headers.Authorization
+import akka.http.scaladsl.model.ContentTypes
 import akka.http.scaladsl.model.StatusCodes._
+import akka.http.scaladsl.model.headers.Authorization
 import akka.http.scaladsl.testkit.ScalatestRouteTest
 import cromiam.server.status.{MockStatusService, StatusService}
 import org.scalatest.{FlatSpec, Matchers}
@@ -23,6 +24,7 @@ class EngineRouteSupportSpec extends FlatSpec with Matchers with ScalatestRouteT
     Get(statsRoutePath).withHeaders(emptyAuthHeader) ~> engineRoutes ~> check {
       status shouldBe Forbidden
       response.entity shouldBe EngineRouteSupport.CromIamStatsForbidden.entity
+      contentType should be(ContentTypes.`text/plain(UTF-8)`)
     }
   }
 
@@ -30,6 +32,7 @@ class EngineRouteSupportSpec extends FlatSpec with Matchers with ScalatestRouteT
     Get(versionRoutePath) ~> engineRoutes ~> check {
       status shouldBe OK
       responseAs[String] shouldBe "Response from Cromwell"
+      contentType should be(ContentTypes.`text/plain(UTF-8)`)
     }
   }
 
@@ -39,6 +42,7 @@ class EngineRouteSupportSpec extends FlatSpec with Matchers with ScalatestRouteT
       assertResult("""{"ok":true,"systems":{"Cromwell":{"ok":true},"Sam":{"ok":true}}}""") {
         responseAs[String]
       }
+      assertResult(ContentTypes.`application/json`)(contentType)
     }
   }
 }