+++ CromIAM/src/test/scala/cromiam/webservice/SubmissionSupportSpec.scala
@@ -5,9 +5,10 @@ import akka.http.scaladsl.model._
 import akka.http.scaladsl.model.headers.{Authorization, OAuth2BearerToken, RawHeader}
 import akka.http.scaladsl.server.AuthorizationFailedRejection
 import akka.http.scaladsl.testkit.{RouteTestTimeout, ScalatestRouteTest}
-import org.scalatest.{FlatSpec, Matchers}
-import scala.concurrent.duration._
 import akka.testkit._
+import org.scalatest.{FlatSpec, Matchers}
+
+import scala.concurrent.duration._
 
 class SubmissionSupportSpec extends FlatSpec with Matchers with ScalatestRouteTest with SubmissionSupport {
   override val cromwellClient = new MockCromwellClient()
@@ -48,12 +49,15 @@ class SubmissionSupportSpec extends FlatSpec with Matchers with ScalatestRouteTe
     Post(submitPath).withHeaders(goodAuthHeaders).withEntity(formData) ~> submitRoute ~> check {
       status shouldEqual StatusCodes.OK
       responseAs[String] shouldBe "Response from Cromwell"
+      contentType should be(ContentTypes.`text/plain(UTF-8)`)
     }
   }
 
   it should "fail with BadRequest for unauthorized SAM user and not forward the request to Cromwell" in {
     Post(submitPath).withHeaders(badAuthHeaders).withEntity(formData) ~> submitRoute ~> check {
       status shouldEqual StatusCodes.BadRequest
+      responseAs[String] should be("Can't register collection with Sam. Status code: 400 Bad Request")
+      contentType should be(ContentTypes.`text/plain(UTF-8)`)
     }
   }